添加了路线最低价格的mapper 增加了element-ui的自定义样式

+++ ruoyi-system/src/main/java/com/ruoyi/travel/mapper/ProductRouteMapper.java
@@ -1,5 +1,6 @@
 package com.ruoyi.travel.mapper;
 
+import java.math.BigDecimal;
 import java.util.List;
 import com.ruoyi.travel.domain.ProductRoute;
 
@@ -41,7 +42,7 @@ public interface ProductRouteMapper
      * @param productId 产品id
      * @return 价格最小值
      */
-    public int getProductRoutePriceMin(Long productId);
+    public BigDecimal getProductRoutePriceMin(Long productId);
 
     /**
      * 新增产品路线
